Address more PR comments
