+++ hi-ml/testhiml/testhiml/test_run_ml.py
@@ -15,7 +15,7 @@ from health_ml.lightning_container import LightningContainer
 from health_ml.run_ml import MLRunner
 from health_ml.utils.common_utils import is_gpu_available
 from health_azure.utils import is_global_rank_zero
-from health_ml.utils.logging import MLFlowLogger
+from health_ml.utils.lightning_loggers import MLFlowLogger
 from testazure.utils_testazure import DEFAULT_WORKSPACE
 from testhiml.utils.fixed_paths_for_tests import mock_run_id
